# openapi.yaml
openapi: 3.0.3
info:
  title: Arsivinyo Memegy API
  version: 0.1.0
  description: |
    A high-performance meme archive backend built with Rust. 
    This API allows users to upload, like, and discover short meme videos.

    ## Authentication
    This API uses **Bearer Token** authentication via JWT. Obtain tokens through the Google OAuth flow.

    ## Rate Limits
    - **Regular Upload**: 256 MB per hour per user
    - **Anonymous Upload**: 128 MB per hour per user (stricter due to abuse risk)
    - **Feed Access**: 30 requests per minute
    - **Video Deletion**: 10 deletions per hour per user
    - **Metadata Updates**: 30 updates per hour per user

  contact:
    name: Arsivinyo Team
  license:
    name: MIT
servers:
  - url: http://localhost:80
    description: Local development server
  - url: https://api.memegy.arsivinyo.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from the authentication flow
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token (valid for 15 minutes)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: Opaque refresh token (valid for 14 days)
          example: a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6
        user:
          $ref: "#/components/schemas/UserDto"
      required:
        - access_token
        - refresh_token
        - user
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: john_doe
        email:
          type: string
          format: email
          example: john@example.com
        avatar_url:
          type: string
          format: uri
          nullable: true
          example: https://lh3.googleusercontent.com/a/ACg8oc...
      required:
        - id
        - username
        - email
...
    UsernameRules:
      type: object
      properties:
        min_length:
          type: integer
          example: 3
        max_length:
          type: integer
          example: 20
        pattern:
          type: string
          example: "^[a-z0-9_]{3,20}$"
      required:
        - min_length
        - max_length
        - pattern
    UsernameRequiredResponse:
      type: object
      properties:
        error:
          type: string
          example: username_required
        signup_ticket:
          type: string
          example: a1b2c3d4e5f6g7h8
        suggested_username:
          type: string
          example: john_doe
        requires_age_confirmation:
          type: boolean
          example: true
        required_terms_version:
          type: string
          example: v1
        terms_url:
          type: string
          nullable: true
          example: https://example.com/terms
        rules:
          $ref: "#/components/schemas/UsernameRules"
      required:
        - error
        - signup_ticket
        - suggested_username
        - requires_age_confirmation
        - required_terms_version
        - terms_url
        - rules
...
    InitUploadRequest:
      type: object
      properties:
        filename:
          type: string
          description: Original filename with extension
          example: funny_cat.mp4
        size_bytes:
          type: integer
          format: int64
          description: File size in bytes
          example: 5242880
          minimum: 1
        is_nsfw:
          type: boolean
          description: Whether the uploaded content is NSFW
          example: false
      required:
        - filename
        - size_bytes
        - is_nsfw
    InitUploadResponse:
      type: object
      properties:
        video_id:
          type: string
          format: uuid
          description: Unique identifier for the video
          example: 550e8400-e29b-41d4-a716-446655440000
        upload_url:
          type: string
          format: uri
          description: Presigned URL for direct upload to storage (valid for 1 hour)
          example: http://localhost:9000/videos-raw/550e8400-e29b-41d4-a716-446655440000/video.mp4?X-Amz-Algorithm=...
      required:
        - video_id
        - upload_url
...
    ModeStatusResponse:
      type: object
      properties:
        enabled:
          type: boolean
      required:
        - enabled
...
  responses:
    UnauthorizedError:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Unauthorized
...
    TooManyRequestsError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Rate limit exceeded
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Internal server error
paths:
...
  /auth/google/login:
    get:
      tags:
        - Authentication
      summary: Initiate Google OAuth2 flow
      description: |
        Redirects the user to Google's OAuth2 consent page.
        After authentication, Google redirects back to `/auth/google/callback`.

        **Note for API Testing:**
        This endpoint is designed for browser-based flows and will redirect to Google's login page.
        For API testing without a browser, use the `/auth/dev/login` endpoint instead (development only).
      operationId: googleLogin
      parameters:
        - name: source
          in: query
          required: false
          description: OAuth flow source (determines PKCE requirements)
          schema:
            type: string
            enum: [web, mobile, popup]
            default: web
        - name: code_challenge
          in: query
          required: false
          description: PKCE code challenge (base64url SHA256). Required for mobile flows. Must be 43 chars.
          schema:
            type: string
            minLength: 43
            maxLength: 43
            pattern: "^[A-Za-z0-9_-]{43}$"
        - name: code_challenge_method
          in: query
          required: false
          description: PKCE method. Only S256 supported.
          schema:
            type: string
            enum: [S256]
      responses:
        "400":
          description: Bad request (mobile without PKCE, invalid challenge)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "302":
          description: Redirect to Google OAuth2 consent page
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: https://accounts.google.com/o/oauth2/v2/auth?client_id=...
  /auth/dev/login:
    post:
      tags:
        - Authentication
      summary: Development login (bypasses Google OAuth)
      description: |
        **⚠️ DEVELOPMENT ONLY** - Only available when `RUST_LOG` contains "debug" or "trace".

        Creates or retrieves a test user and issues access/refresh tokens without Google OAuth.
        Perfect for testing the API with Postman, curl, or other API clients.

        **Production Behavior:**
        This endpoint returns 404 in production environments.
      operationId: devLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: test_user
                email:
                  type: string
                  format: email
                  example: test@example.com
              required:
                - username
                - email
      responses:
        "200":
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "404":
          description: Endpoint not available (production mode)
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/google/callback:
    get:
      tags:
        - Authentication
      summary: Google OAuth2 callback
      description: |
        Handles the OAuth2 callback from Google.
        Exchanges the authorization code for user information and issues access/refresh tokens.
      operationId: googleCallback
      parameters:
        - name: code
          in: query
          required: true
          description: Authorization code from Google
          schema:
            type: string
            example: 4/0AY0e-g7...
        - name: state
          in: query
          required: true
          description: CSRF protection token
          schema:
            type: string
        - name: code_verifier
          in: query
          required: false
          description: PKCE code verifier. Required if PKCE was used in login flow.
          schema:
            type: string
            minLength: 43
            maxLength: 128
            pattern: '^[A-Za-z0-9_\-.~]+$'
      responses:
        "302":
          description: Redirect to frontend with OTC
        "400":
          description: Invalid state or PKCE validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /auth/exchange-otc:
    post:
      tags:
        - Authentication
      summary: Exchange one-time code for tokens
      description: |
        Exchanges OTC from OAuth callback for access/refresh tokens.
        Rate limited to 5 attempts per IP per minute (configurable).
      operationId: exchangeOtc
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  example: "a1b2c3d4e5f6g7h8"
      responses:
        "200":
          description: Successfully exchanged OTC
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Username setup required before issuing final tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameRequiredResponse"
        "400":
          description: Invalid or expired OTC
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /videos/init:
    post:
      tags:
        - Videos
      summary: Initialize video upload
      description: |
        Request a presigned URL for uploading a video file.
        New uploads must explicitly provide `is_nsfw`.

        **Upload Flow:**
        1. Call this endpoint to get a `video_id` and `upload_url`
        2. Upload the file directly to `upload_url` using HTTP PUT
        3. Call `/videos/{id}/confirm` to trigger processing

        **Rate Limiting:**
        - Max 50MB per hour per user (configurable)
        - Limit is based on cumulative file size, not number of uploads
      operationId: initUpload
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitUploadRequest"
      responses:
        "200":
          description: Upload initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitUploadResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "429":
          $ref: "#/components/responses/TooManyRequestsError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /videos/init/anonymous:
    post:
      tags:
        - Videos
      summary: Initialize anonymous video upload
      description: |
        Request a presigned URL for uploading an anonymous video file.
        Anonymous videos do not expose the uploader's identity in the feed.
        New uploads must explicitly provide `is_nsfw`.

        **Differences from Regular Upload:**
        - Stricter rate limit: 128 MB/hour (50% of regular limit)
        - No uploader info shown in feed responses
        - User still owns the video (can delete/update via ownership)

        **Upload Flow:**
        1. Call this endpoint to get a `video_id` and `upload_url`
        2. Upload the file directly to `upload_url` using HTTP PUT
        3. Call `/videos/{id}/confirm` to trigger processing

        **Rate Limiting:**
        - Max 128 MB per hour per user (half of regular limit)
        - Limit is based on cumulative file size
      operationId: initAnonymousUpload
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitUploadRequest"
      responses:
        "200":
          description: Anonymous upload initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitUploadResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "429":
          $ref: "#/components/responses/TooManyRequestsError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /system/maintenance:
    get:
      tags:
        - System
      summary: Get maintenance mode status
      operationId: getMaintenanceStatus
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Maintenance mode status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModeStatusResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "429":
          $ref: "#/components/responses/TooManyRequestsError"
        "503":
          description: Status check temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...